Add check for whether dirs already exist

+++ paper_experiments.sh
@@ -3,83 +3,100 @@
 mkdir mse-homolog-loss
 
 # MSE homolog loss
+if [ ! -d mse-homolog-loss/mse_b512_e20 ]; then
 echo 'MSE homolog loss with batch_size 512'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss mse \
-    --model graphsage_maxpool --max_total_steps 100000000000 --epochs 20 --validate_iter 10 \
-    --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid
+    --model graphsage_maxpool --max_total_steps 100000000000 --epochs 10 --validate_iter 10 \
+    --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05
 mkdir mse-homolog-loss/mse_b512_e20
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy mse-homolog-loss/mse_b512_e20/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to mse-homolog-loss/mse_b512_e20'
+fi
 
+if [ ! -d mse-homolog-loss/mse_b256_e20 ]; then
 echo 'MSE homolog loss with batch_size 256'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss mse \
-    --model graphsage_maxpool --batch_size 256 --max_total_steps 100000000000 --epochs 20 \
-    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid
+    --model graphsage_maxpool --batch_size 256 --max_total_steps 100000000000 --epochs 10 \
+    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05
 mkdir mse-homolog-loss/mse_b256_e20
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy mse-homolog-loss/mse_b256_e20/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to mse-homolog-loss/mse_b256_e20'
+fi
 
 
+if [ ! -d cross-homolog-loss ]; then
 mkdir cross-homolog-loss
+fi
 
 # dot product homolog loss
+if [ ! -d cross-homolog-loss/cross_b512_e20 ]; then
 echo 'cross product homolog loss with batch_size 512'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss cross \
-    --model graphsage_maxpool --max_total_steps 100000000000 --epochs 20 --validate_iter 10 \
-    --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid
+    --model graphsage_maxpool --max_total_steps 100000000000 --epochs 10 --validate_iter 10 \
+    --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05
 mkdir cross-homolog-loss/cross_b512_e20
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy cross-homolog-loss/cross_b512_e20/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to cross-homolog-loss/cross_b512_e20'
+fi
 
 
 # dot product homolog loss
+if [ ! -d cross-homolog-loss/cross_b256_e20 ]; then
 echo 'cross product homolog loss with batch_size 256'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss cross \
-    --model graphsage_maxpool --batch_size 256 --max_total_steps 100000000000 --epochs 20 \
-    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid
+    --model graphsage_maxpool --batch_size 256 --max_total_steps 100000000000 --epochs 10 \
+    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05
 mkdir cross-homolog-loss/cross_b256_e20
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy cross-homolog-loss/cross_b256_e20/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to cross-homolog-loss/cross_b256_e20'
+fi
 
-
+if [ ! -d dot-homolog-loss ]; then
 mkdir dot-homolog-loss
+fi
 
 # dot product homolog loss
+if [ ! -d dot-homolog-loss/dot_b512_e20 ]; then
 echo 'dot product homolog loss with batch_size 512'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss dot \
-    --model graphsage_maxpool --max_total_steps 100000000000 --epochs 20 --validate_iter 10 \
-    --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid
+    --model graphsage_maxpool --max_total_steps 100000000000 --epochs 10 --validate_iter 10 \
+    --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05
 mkdir dot-homolog-loss/dot_b512_e20
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy dot-homolog-loss/dot_b512_e20/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to dot-homolog-loss/dot_b512_e20'
+fi
 
 
 # dot product homolog loss
+if [ ! -d dot-homolog-loss/dot_b256_e20 ]; then
 echo 'dot product homolog loss with batch_size 256'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss dot \
-    --model graphsage_maxpool --batch_size 256 --max_total_steps 100000000000 --epochs 20 \
-    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid
+    --model graphsage_maxpool --batch_size 256 --max_total_steps 100000000000 --epochs 10 \
+    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05
 mkdir dot-homolog-loss/dot_b256_e20
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy dot-homolog-loss/dot_b256_e20/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to dot-homolog-loss/dot_b256_e20'
+fi
 
 
 # dot product homolog loss
+if [ ! -d dot-homolog-loss/dot_b512_e20_hi1000 ]; then
 echo 'dot product homolog loss with homolog importance set to 1000'
 python -m graphsage.unsupervised_train --train_prefix ./cge/ppi --homolog_loss dot \
-    --model graphsage_maxpool --batch_size 512 --max_total_steps 100000000000 --epochs 20 \
-    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 --sigmoid \
+    --model graphsage_maxpool --batch_size 512 --max_total_steps 100000000000 --epochs 10 \
+    --validate_iter 10 --identity_dim 128 --dim_1 400 --dim_2 400 --dropout 0.05 \
     --homolog_importance 1000
 mkdir dot-homolog-loss/dot_b512_e20_hi1000
 mv unsup-cge/graphsage_maxpool_small_0.000010/val.npy dot-homolog-loss/dot_b512_e20_hi1000/val.npy
 rm -rf unsup-cge
 echo 'Embeddings have been saved to dot-homolog-loss/dot_b512_e20_hi1000'
+fi
 
 
 ##### SPLIT EMBEDDINGS PER ORGANISM #####
